checkpoint test refactor
